% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaf_dimensions.R
\name{extract_calibration_square}
\alias{extract_calibration_square}
\title{Extract the calibration square as a pixset.}
\usage{
extract_calibration_square(im, rmin, gmax, bmax)
}
\arguments{
\item{im}{a `cimg` object.}

\item{rmin}{minimum value of red spectrum (on scale of 0 to 1) for
detecting the calibration square. Default = 0.8.}

\item{gmax}{maximum value of green spectrum (on scale of 0 to 1) for
detecting the calibration square. Default = 0.6.}

\item{bmax}{maximum value of blue spectrum (on scale of 0 to 1) for
detecting the calibration square.  Default = 0.6.}
}
\value{
a one-dimensional pixset corresponding to the calibration square.
}
\description{
Isolates the area of an image containing a calibration square.
}
\details{
The calibration square is detected as being the region with pixels such that
the RGB color value has an R value greater than `rmin`, a G value less than
`gmax`, and a B value less than `bmax`.
}
