% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaf_dimensions.R
\name{pixset_dimensions}
\alias{pixset_dimensions}
\title{Compute the length, width, and area of a pixset.}
\usage{
pixset_dimensions(px)
}
\arguments{
\item{px}{one-dimensional pixset.}
}
\value{
a named vector of length 3 containing the length, width, and area.
}
\description{
This computes the tallest column (length) of pixels in a pixset, the widest
row (width) of pixels in a pixset, and the total number of pixels (area) of
a pixset. Gaps are overlooked when counting the length and width.
}
