% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_selection.R
\name{plot_selection}
\alias{plot_selection}
\title{Visualize selection.}
\usage{
plot_selection(im, region, color = "green")
}
\arguments{
\item{im}{a `cimg` object.}

\item{region}{pixset representing the region selected.}

\item{color}{color to highlight (default = 'green').}
}
\value{
a plot object.
}
\description{
Highlights a selection (like the calibration square or the leaf) on a
grayscale version of the image. This is a quick way of verifying the settings
for isolating the region.
}
