% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgname.R
\docType{package}
\name{rhleaves-package}
\alias{rhleaves-package}
\title{Measurements on leaves taken from the living walls on the Rose-Hulman
Institute of Technology campus.}
\description{
[Rose-Hulman Institute of Technology](https://www.rose-hulman.edu/) maintains
two "living walls" on campus.  Each living wall is comprised of a mosaic of
six species of plants.  Each of the plant has almond-shaped leaves.  A sample
of leaves was taken in April 2025.  This package contains measurements taken
on the leaves, an image of each leaf, and a function for computing the area
of a leaf from an image.
}
\details{
The dataset is called `rhleaves` and the function for computing the
dimensions of a leaf from its image is `leaf_dimensions()`.
}
\author{
\strong{Maintainer}: Eric Reyes \email{reyesem@rose-hulman.edu}

}
